+++ openhands/runtime/impl/eventstream/eventstream_runtime.py
@@ -148,6 +148,9 @@ class EventStreamRuntime(ActionExecutionClient):
                 f'Installing extra user-provided dependencies in the runtime image: {self.config.sandbox.runtime_extra_deps}',
             )
 
+    async def _get_api_url(self):
+        return self.api_url
+
     async def connect(self):
         self.send_status_message('STATUS$STARTING_RUNTIME')
         try:
